% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{kipp_colors}
\alias{kipp_colors}
\title{KIPP Colors}
\format{An object of class \code{list} of length 14.}
\usage{
kipp_colors
}
\description{
Named list of colors from the [KIPP official brand guide](www.kipp.org/brand).
}
\details{
The list of available colors is: darkblue, blue, lightblue, skyblue, darkorange, orange,
lightorange, yellow, green, lightgreen, darkgray, gray, lightgray, and brown
}
\keyword{datasets}
