% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{kipp_palettes}
\alias{kipp_palettes}
\title{KIPP Color Palettes}
\format{An object of class \code{list} of length 10.}
\usage{
kipp_palettes
}
\description{
A collection of color palettes based on the [KIPP official brand guide](www.kipp.org/brand).
}
\details{
The list of available palettes is:
* kipp_qual
* kipp_div
* kipp_greenorgange
* kipp_blueorange
* kipp_bluegray
* kipp_greengray
* kipp_orangegray
* kipp_greenbrown
* kipp_browngray
* kipp_grays
}
\examples{

# Make an x-y plot using the kipp_div palette
library(tidyverse)
df <- data.frame(x = rnorm(100, 0, 20),
          y = rnorm(100, 0, 20),
          cl = sample(letters[1:8], 100, replace=TRUE))
ggplot(df, aes(x, y, colour=cl, shape=cl)) +
  geom_point(size=4) +
  scale_colour_kipp() +
  theme_bw() +
  theme(aspect.ratio=1)

# Make a histogram using the kipp_qual palette
ggplot(df, aes(x, fill=cl)) + geom_histogram() +
  scale_fill_kipp(palette="kipp_qual")
}
\keyword{datasets}
