% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_kipp}
\alias{scale_color_kipp}
\alias{scale_colour_kipp}
\title{Setup color palette for ggplot2}
\usage{
scale_color_kipp(palette = "kipp_div", discrete = TRUE, alpha = 1,
  reverse = FALSE, ...)

scale_colour_kipp(palette = "kipp_div", discrete = TRUE, alpha = 1,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Choose from 'kipp_palettes' list}

\item{discrete}{whether to use a discrete color palette}

\item{alpha}{transparency}

\item{reverse}{logical, Reverse the order of the colours?}

\item{...}{additional arguments to pass to scale_color_gradientn}
}
\description{
Setup color palette for ggplot2
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = factor(cyl))) +
  scale_colour_kipp(palette="kipp_qual")
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = hp)) +
  scale_colour_kipp(palette="kipp_qual", discrete = FALSE)
ggplot(data = mpg) +
  geom_point(mapping = aes(x = displ, y = hwy, color = class)) +
  scale_colour_kipp(palette="kipp_bluegray")
ggplot(diamonds) + geom_bar(aes(x = cut, fill = clarity)) +
  scale_fill_kipp()
}
