% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_kipp_light}
\alias{theme_kipp_light}
\title{Theme inspired by kipp.org/brand}
\usage{
theme_kipp_light(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Theme inspired by the plots on
\href{KIPP Brand Guidlines}{http://www.kipp.org/brand}.
}
\concept{themes kipp}
