% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_kipp_min}
\alias{theme_kipp_min}
\title{A minimal Theme inspired by kipp.org/brand}
\usage{
theme_kipp_min()
}
\description{
Theme inspired by the plots on
\href{KIPP Brand Guidlines}{http://www.kipp.org/brand}.
}
\seealso{
Other themes kipp: \code{\link{theme_kipp_light}}
}
\concept{themes kipp}
