% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{use_kipp_chicago_css}
\alias{use_kipp_chicago_css}
\title{Creates www/kippcolors_*.css for using KIPP colors with `shinydashboard` apps}
\usage{
use_kipp_chicago_css(color = "blue", directory = ".")
}
\arguments{
\item{color}{name header menu main color}

\item{directory}{locatoin where www/kippchicago_*.css should be stored}
}
\description{
Creates www/kippcolors_*.css for using KIPP colors with `shinydashboard` apps
}
