% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kippcolors-package.R
\docType{package}
\name{kippcolors-package}
\alias{kippcolors}
\alias{kippcolors-package}
\title{kippcolors: KIPP compliant colors from [brand guidelines](http://www.kipp.org/brand)}
\description{
More about what it does (maybe more than one line)
Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
